flask_app/datalake_from_scratch/requirements.txt
autopep8==2.0.1
blinker==1.6.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.3
flake8==6.1.0
Flask==2.3.2
Flask-HTTPAuth==4.8.0
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==0.4.3
numpy==1.26.4
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.11.0
pyflakes==3.1.0
requests==2.31.0
rsa==4.9
urllib3==2.2.1
Werkzeug==2.3.7
